+++ src/extension/exponent/xdlInterface.ts
@@ -1,62 +1,56 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
-import { CommandExecutor, CommandVerbosity } from "../../common/commandExecutor";
-import { HostPlatform } from "../../common/hostPlatform";
-import { OutputChannelLogger } from "../log/OutputChannelLogger";
-
 import * as XDLPackage from "xdl";
-import * as path from "path";
-import { findFileInFolderHierarchy } from "../../common/extensionHelper";
-import customRequire from "../../common/customRequire";
+import * as MetroConfigPackage from "metro-config";
+import { PackageLoader, PackageConfig } from "../../common/packageLoader";
+import { SettingsHelper } from "../settingsHelper";
 
-const logger: OutputChannelLogger = OutputChannelLogger.getMainChannel();
+const XDL_PACKAGE = "xdl";
+const METRO_CONFIG_PACKAGE = "@expo/metro-config";
 
-const XDL_PACKAGE = "@expo/xdl";
-const EXPO_DEPS: string[] = [
+const xdlPackageConfig = new PackageConfig(
     XDL_PACKAGE,
-    "@expo/ngrok", // devDependencies for xdl
-];
+    SettingsHelper.getExpoDependencyVersion(XDL_PACKAGE),
+);
+const metroConfigPackageConfig = new PackageConfig(
+    METRO_CONFIG_PACKAGE,
+    SettingsHelper.getExpoDependencyVersion("metroConfig"),
+);
 
-let xdlPackage: Promise<typeof XDLPackage>;
+const ngrokPackageConfig = new PackageConfig(
+    xdlPackageConfig.getPackageName(),
+    xdlPackageConfig.getVersion(),
+    "build/start/resolveNgrok",
+);
 
-function getPackage(): Promise<typeof XDLPackage> {
-    if (xdlPackage) {
-        return xdlPackage;
-    }
-    // Don't do the require if we don't actually need it
-    try {
-        logger.debug("Getting exponent dependency.");
-        const xdl = customRequire(XDL_PACKAGE);
-        xdlPackage = Promise.resolve(xdl);
-        return xdlPackage;
-    } catch (e) {
-        if (e.code === "MODULE_NOT_FOUND") {
-            logger.debug("Dependency not present. Installing it...");
-        } else {
-            throw e;
-        }
-    }
-    let commandExecutor = new CommandExecutor(
-        path.dirname(findFileInFolderHierarchy(__dirname, "package.json") || __dirname),
-        logger,
-    );
-    xdlPackage = commandExecutor
-        .spawnWithProgress(
-            HostPlatform.getNpmCliCommand("npm"),
-            ["install", ...EXPO_DEPS, "--verbose", "--no-save"],
-            { verbosity: CommandVerbosity.PROGRESS },
-        )
-        .then((): typeof XDLPackage => {
-            return customRequire(XDL_PACKAGE);
-        });
-    return xdlPackage;
-}
+// There is the problem with '--no-save' flag for 'npm install' command for npm v6.
+// Installing npm dependencies with the `--no-save` flag will remove
+// other dependencies that were installed previously in the same manner (https://github.com/npm/cli/issues/1460).
+// So we should workaround it passing all packages for install to only one npm install command
+const EXPO_DEPS: PackageConfig[] = [xdlPackageConfig, metroConfigPackageConfig];
+
+export const getXDLPackage: () => Promise<
+    typeof XDLPackage
+> = PackageLoader.getInstance().generateGetPackageFunction<typeof XDLPackage>(
+    xdlPackageConfig,
+    ...EXPO_DEPS,
+);
+export const getMetroConfigPackage: () => Promise<
+    typeof MetroConfigPackage
+> = PackageLoader.getInstance().generateGetPackageFunction<typeof MetroConfigPackage>(
+    metroConfigPackageConfig,
+    ...EXPO_DEPS,
+);
+export const getNgrokResolver: () => Promise<XDLPackage.ResolveNgrok> = PackageLoader.getInstance().generateGetPackageFunction<XDLPackage.ResolveNgrok>(
+    ngrokPackageConfig,
+    ...EXPO_DEPS,
+);
 
 export type IUser = XDLPackage.IUser;
 
-export function configReactNativeVersionWargnings(): Promise<void> {
-    return getPackage().then(xdl => {
+export function configReactNativeVersionWarnings(): Promise<void> {
+    return getXDLPackage().then(xdl => {
         xdl.Config.validation.reactNativeVersionWarnings = false;
     });
 }
@@ -65,64 +59,84 @@ export function attachLoggerStream(
     rootPath: string,
     options?: XDLPackage.IBunyanStream | any,
 ): Promise<void> {
-    return getPackage().then(xdl => xdl.ProjectUtils.attachLoggerStream(rootPath, options));
-}
-
-export function supportedVersions(): Promise<string[]> {
-    return getPackage().then(xdl => xdl.Versions.facebookReactNativeVersionsAsync());
+    return getXDLPackage().then(xdl => xdl.ProjectUtils.attachLoggerStream(rootPath, options));
 }
 
 export function currentUser(): Promise<XDLPackage.IUser> {
-    return getPackage().then(xdl =>
+    return getXDLPackage().then(xdl =>
         xdl.User ? xdl.User.getCurrentUserAsync() : xdl.UserManager.getCurrentUserAsync(),
     );
 }
 
 export function login(username: string, password: string): Promise<XDLPackage.IUser> {
-    return getPackage().then(xdl =>
+    return getXDLPackage().then(xdl =>
         xdl.User
             ? xdl.User.loginAsync("user-pass", { username: username, password: password })
-            : xdl.UserManager.loginAsync("user-pass", { username: username, password: password }),
+            : xdl.UserManager.loginAsync("user-pass", {
+                  username: username,
+                  password: password,
+              }),
     );
 }
 
-export function mapVersion(reactNativeVersion: string): Promise<string> {
-    return getPackage().then(xdl =>
-        xdl.Versions.facebookReactNativeVersionToExpoVersionAsync(reactNativeVersion),
-    );
+export function getExpoSdkVersions(): Promise<XDLPackage.SDKVersions> {
+    return getXDLPackage().then(xdl => xdl.Versions.sdkVersionsAsync());
+}
+
+export function getReleasedExpoSdkVersions(): Promise<XDLPackage.SDKVersions> {
+    return getXDLPackage().then(xdl => xdl.Versions.releasedSdkVersionsAsync());
 }
 
 export function publish(
     projectRoot: string,
     options?: XDLPackage.IPublishOptions,
 ): Promise<XDLPackage.IPublishResponse> {
-    return getPackage().then(xdl => xdl.Project.publishAsync(projectRoot, options));
+    return getXDLPackage().then(xdl => xdl.Project.publishAsync(projectRoot, options));
 }
 
 export function setOptions(projectRoot: string, options?: XDLPackage.IOptions): Promise<void> {
-    return getPackage().then(xdl => xdl.Project.setOptionsAsync(projectRoot, options));
+    return getXDLPackage().then(xdl => xdl.Project.setOptionsAsync(projectRoot, options));
 }
 
 export function startExponentServer(projectRoot: string): Promise<void> {
-    return getPackage().then(xdl => xdl.Project.startExpoServerAsync(projectRoot));
+    return getXDLPackage().then(xdl => xdl.Project.startExpoServerAsync(projectRoot));
 }
 
 export function startTunnels(projectRoot: string): Promise<void> {
-    return getPackage().then(xdl => xdl.Project.startTunnelsAsync(projectRoot));
+    return getXDLPackage().then(xdl => xdl.Project.startTunnelsAsync(projectRoot));
 }
 
 export function getUrl(projectRoot: string, options?: XDLPackage.IUrlOptions): Promise<string> {
-    return getPackage().then(xdl => xdl.UrlUtils.constructManifestUrlAsync(projectRoot, options));
+    return getXDLPackage().then(xdl =>
+        xdl.UrlUtils.constructManifestUrlAsync(projectRoot, options),
+    );
 }
 
 export function stopAll(projectRoot: string): Promise<void> {
-    return getPackage().then(xdl => xdl.Project.stopAsync(projectRoot));
+    return getXDLPackage().then(xdl => xdl.Project.stopAsync(projectRoot));
 }
 
 export function startAdbReverse(projectRoot: string): Promise<boolean> {
-    return getPackage().then(xdl => xdl.Android.startAdbReverseAsync(projectRoot));
+    return getXDLPackage().then(xdl => xdl.Android.startAdbReverseAsync(projectRoot));
 }
 
 export function stopAdbReverse(projectRoot: string): Promise<void> {
-    return getPackage().then(xdl => xdl.Android.stopAdbReverseAsync(projectRoot));
+    return getXDLPackage().then(xdl => xdl.Android.stopAdbReverseAsync(projectRoot));
+}
+
+export function getMetroConfig(projectRoot: string): Promise<MetroConfigPackage.IMetroConfig> {
+    return getMetroConfigPackage().then(metroConfigPackage =>
+        metroConfigPackage.loadAsync(projectRoot),
+    );
+}
+
+export function isNgrokInstalled(projectRoot: string): Promise<boolean> {
+    return getNgrokResolver()
+        .then(ngrokResolver =>
+            ngrokResolver.resolveNgrokAsync(projectRoot, {
+                shouldPrompt: false,
+                autoInstall: false,
+            }),
+        )
+        .then(ngrok => !!ngrok);
 }