+++ src/extension/windows/windowsPlatform.ts
@@ -8,6 +8,7 @@ import { OutputVerifier, PatternToFailure } from "../../common/outputVerifier";
 import { TelemetryHelper } from "../../common/telemetryHelper";
 import { CommandExecutor } from "../../common/commandExecutor";
 import { InternalErrorCode } from "../../common/error/internalErrorCode";
+import { AppLauncher } from "../appLauncher";
 
 /**
  * Windows specific platform implementation for debugging RN applications.
@@ -21,7 +22,19 @@ export class WindowsPlatform extends GeneralMobilePlatform {
             pattern: "Unrecognized command 'run-windows'",
             errorCode: InternalErrorCode.WinRNMPPluginIsNotInstalled,
         },
+        {
+            pattern: /×.+$/gm,
+            errorCode: InternalErrorCode.WinRunCommandFailed,
+        },
     ];
+
+    public reloadApp(appLauncher: AppLauncher): Promise<void> {
+        const worker = appLauncher.getAppWorker();
+        if (worker) {
+            worker.reloadAppCommand();
+        }
+        return Promise.resolve();
+    }
 
     constructor(protected runOptions: IWindowsRunOptions, platformDeps: MobilePlatformDeps = {}) {
         super(runOptions, platformDeps);
@@ -70,6 +83,7 @@ export class WindowsPlatform extends GeneralMobilePlatform {
             }
 
             const runWindowsSpawn = new CommandExecutor(
+                this.runOptions.nodeModulesRoot,
                 this.projectPath,
                 this.logger,
             ).spawnReactCommand(`run-${this.platformName}`, this.runArguments, { env });