+++ test/smoke/package/src/expoDebug.test.ts
@@ -6,10 +6,13 @@ import * as path from "path";
 import { Application } from "../../automation";
 import { AppiumClient, AppiumHelper, Platform } from "./helpers/appiumHelper";
 import { SmokeTestsConstants } from "./helpers/smokeTestsConstants";
-import { findFile, findStringInFile, sleep, waitUntil } from "./helpers/utilities";
+import { findFile, findStringInFile, isLoggedInExpo, sleep, waitUntil } from "./helpers/utilities";
 import { androidEmulatorManager, iosSimulatorManager, vscodeManager } from "./main";
 import { TestRunArguments } from "./helpers/testConfigProcessor";
 import { SmokeTestLogger } from "./helpers/smokeTestLogger";
+import TestProject from "./helpers/testProject";
+import AutomationHelper from "./helpers/AutomationHelper";
+import AndroidEmulatorManager from "./helpers/androidEmulatorManager";
 
 const EXPO_APP_LAUNCH_TIMEOUT = 120_000;
 const ExpoSuccessPattern = "Tunnel ready";
@@ -17,13 +20,15 @@ const ExpoFailurePattern = "XDLError";
 
 const EXPO_APP_PACKAGE_NAME = SmokeTestsConstants.expoPackageName;
 const EXPO_APP_ACTIVITY_NAME = `${SmokeTestsConstants.expoPackageName}.experience.HomeActivity`;
-const ExpoDebugConfigName = "Debug in Exponent";
+const ExpoTunnelDebugConfigName = "Debug in Exponent (Tunnel)";
 const ExpoLanDebugConfigName = "Debug in Exponent (LAN)";
 const ExpoLocalDebugConfigName = "Debug in Exponent (Local)";
 
 const ExpoSetBreakpointOnLine = 1;
 // Time for Android Expo Debug Test before it reaches timeout
 const debugExpoTestTime = SmokeTestsConstants.expoTestTimeout;
+const expoLogginErrorMessage =
+    "It seems you are not currently logged into Expo account. To successfully pass this test, you must be logged into Expo account";
 
 interface ExpoLaunch {
     successful: boolean;
@@ -31,26 +36,50 @@ interface ExpoLaunch {
 }
 
 export function startExpoTests(
-    expoWorkspace: string,
-    pureWorkspace: string,
+    expoProject: TestProject,
+    pureProject: TestProject,
     testParameters: TestRunArguments,
 ): void {
     describe("Expo tests", () => {
         let app: Application;
         let expoFirstLaunch = true;
         let client: AppiumClient;
+        let automationHelper: AutomationHelper;
+
+        async function initApp(
+            workspaceOrFolder: string,
+            sessionName?: string,
+            locale?: string,
+        ): Promise<Application> {
+            app = await vscodeManager.runVSCode(workspaceOrFolder, sessionName, locale);
+            automationHelper = new AutomationHelper(app);
+            return app;
+        }
 
         async function disposeAll() {
-            SmokeTestLogger.info("Dispose all ...");
-            if (app) {
-                SmokeTestLogger.info("Stopping React Native packager ...");
-                await app.workbench.quickaccess.runCommand(SmokeTestsConstants.stopPackagerCommand);
-                await sleep(3000);
-                await app.stop();
-            }
-            if (client) {
-                await client.closeApp();
-                await client.deleteSession();
+            try {
+                SmokeTestLogger.info("Dispose all ...");
+                if (app) {
+                    SmokeTestLogger.info("Stopping React Native packager ...");
+                    await automationHelper.runCommandWithRetry(
+                        SmokeTestsConstants.stopPackagerCommand,
+                    );
+                    await sleep(3000);
+                    SmokeTestLogger.info("Stopping application ...");
+                    await app.stop();
+                }
+                if (client) {
+                    SmokeTestLogger.info("Closing application ...");
+                    await client.closeApp();
+                    SmokeTestLogger.info("Deleting session ...");
+                    await client.deleteSession();
+                }
+                SmokeTestLogger.info("Clearing android application ...");
+                AndroidEmulatorManager.closeApp(EXPO_APP_PACKAGE_NAME);
+            } catch (error) {
+                SmokeTestLogger.error("Error while disposeAll:");
+                SmokeTestLogger.error(error);
+                // throw error;
             }
         }
 
@@ -102,18 +131,19 @@ export function startExpoTests(
         async function runExpoDebugScenario(
             logFilePath: string,
             testName: string,
-            workspacePath: string,
             debugConfigName: string,
             triesToLaunchApp: number,
         ) {
             SmokeTestLogger.info(`${testName}: Starting debugging`);
             // Scan logs only if launch retries provided (Expo Tunnel scenarios)
             if (triesToLaunchApp <= 1) {
-                await app.workbench.quickaccess.runDebugScenario(debugConfigName);
+                await automationHelper.runDebugScenarioWithRetry(debugConfigName);
+                await automationHelper.runCommandWithRetry("Output: Focus on Output View");
             } else {
                 for (let retry = 1; retry <= triesToLaunchApp; retry++) {
                     let expoLaunchStatus: ExpoLaunch;
-                    await app.workbench.quickaccess.runDebugScenario(debugConfigName);
+                    await automationHelper.runDebugScenarioWithRetry(debugConfigName);
+                    await automationHelper.runCommandWithRetry("Output: Focus on Output View");
                     expoLaunchStatus = await findExpoSuccessAndFailurePatterns(
                         logFilePath,
                         ExpoSuccessPattern,
@@ -126,25 +156,27 @@ export function startExpoTests(
                             assert.fail(`App start has failed after ${retry} retries`);
                         }
                         SmokeTestLogger.warn(`Attempt to start #${retry} failed, retrying...`);
+                        await automationHelper.stopDebuggingWithRetry();
                     }
                 }
             }
         }
 
         async function expoTest(
-            appFileName: string,
+            project: TestProject,
             testName: string,
-            workspacePath: string,
             debugConfigName: string,
             platform: Platform.AndroidExpo | Platform.iOSExpo,
             triesToLaunchApp: number,
         ) {
             let logFilePath = "";
-            app = await vscodeManager.runVSCode(workspacePath, testName);
-            SmokeTestLogger.info(`${testName}: ${workspacePath} directory is opened in VS Code`);
-            await app.workbench.quickaccess.openFile(appFileName);
+            app = await initApp(project.workspaceDirectory, testName);
+            SmokeTestLogger.info(
+                `${testName}: ${project.workspaceDirectory} directory is opened in VS Code`,
+            );
+            await automationHelper.openFileWithRetry(project.projectEntryPointFile);
             await app.workbench.editors.scrollTop();
-            SmokeTestLogger.info(`${testName}: ${appFileName} file is opened`);
+            SmokeTestLogger.info(`${testName}: ${project.projectEntryPointFile} file is opened`);
             await app.workbench.debug.setBreakpointOnLine(ExpoSetBreakpointOnLine);
             SmokeTestLogger.info(
                 `${testName}: Breakpoint is set on line ${ExpoSetBreakpointOnLine}`,
@@ -158,13 +190,7 @@ export function startExpoTests(
             } else {
                 assert.fail("REACT_NATIVE_TOOLS_LOGS_DIR is not defined");
             }
-            await runExpoDebugScenario(
-                logFilePath,
-                testName,
-                workspacePath,
-                debugConfigName,
-                triesToLaunchApp,
-            );
+            await runExpoDebugScenario(logFilePath, testName, debugConfigName, triesToLaunchApp);
 
             await app.workbench.editors.waitForTab("Expo QR Code", false, true);
             await app.workbench.editors.waitForActiveTab("Expo QR Code", false, true);
@@ -191,7 +217,7 @@ export function startExpoTests(
                     Platform.iOS,
                     client,
                     expoURL,
-                    workspacePath,
+                    project.workspaceDirectory,
                     expoFirstLaunch,
                 );
                 expoFirstLaunch = false;
@@ -217,7 +243,7 @@ export function startExpoTests(
                     Platform.Android,
                     client,
                     expoURL,
-                    workspacePath,
+                    project.workspaceDirectory,
                 );
                 // TODO Add listener to trigger that child expo app has been ran instead of using timeout
                 SmokeTestLogger.info(
@@ -231,9 +257,11 @@ export function startExpoTests(
             }
 
             SmokeTestLogger.info(`${testName}: Debugging started`);
-            await app.workbench.debug.waitForStackFrame(
-                sf => sf.name === appFileName && sf.lineNumber === ExpoSetBreakpointOnLine,
-                `looking for ${appFileName} and line ${ExpoSetBreakpointOnLine}`,
+            await automationHelper.waitForStackFrameWithRetry(
+                sf =>
+                    sf.name === project.projectEntryPointFile &&
+                    sf.lineNumber === ExpoSetBreakpointOnLine,
+                `looking for ${project.projectEntryPointFile} and line ${ExpoSetBreakpointOnLine}`,
             );
             SmokeTestLogger.info(`${testName}: Stack frame found`);
             await app.workbench.debug.stepOver();
@@ -242,6 +270,7 @@ export function startExpoTests(
             SmokeTestLogger.info(
                 `${testName}: Searching for \"Test output from debuggee\" string in console`,
             );
+            await automationHelper.runCommandWithRetry("Debug: Focus on Debug Console View");
             let found = await app.workbench.debug.waitForOutput(output =>
                 output.some(line => line.indexOf("Test output from debuggee") >= 0),
             );
@@ -251,29 +280,16 @@ export function startExpoTests(
                 '"Test output from debuggee" string is missing in debug console',
             );
             SmokeTestLogger.success(`${testName}: \"Test output from debuggee\" string is found`);
-            await app.workbench.debug.disconnectFromDebugger();
+            await automationHelper.disconnectFromDebuggerWithRetry();
             SmokeTestLogger.info(`${testName}: Debugging is stopped`);
         }
 
         if (testParameters.RunAndroidTests) {
-            it("Android Expo app Debug test(Tunnel)", async function () {
-                this.timeout(debugExpoTestTime);
-                await expoTest(
-                    "App.tsx",
-                    "Android Expo Debug test(Tunnel)",
-                    expoWorkspace,
-                    ExpoDebugConfigName,
-                    Platform.AndroidExpo,
-                    5,
-                );
-            });
-
             it("Android Pure RN app Expo test(LAN)", async function () {
                 this.timeout(debugExpoTestTime);
                 await expoTest(
-                    "App.js",
+                    pureProject,
                     "Android pure RN Expo test(LAN)",
-                    pureWorkspace,
                     ExpoLanDebugConfigName,
                     Platform.AndroidExpo,
                     1,
@@ -283,9 +299,8 @@ export function startExpoTests(
             it("Android Expo app Debug test(LAN)", async function () {
                 this.timeout(debugExpoTestTime);
                 await expoTest(
-                    "App.tsx",
+                    expoProject,
                     "Android Expo Debug test(LAN)",
-                    expoWorkspace,
                     ExpoLanDebugConfigName,
                     Platform.AndroidExpo,
                     1,
@@ -295,34 +310,35 @@ export function startExpoTests(
             it("Android Expo app Debug test(localhost)", async function () {
                 this.timeout(debugExpoTestTime);
                 await expoTest(
-                    "App.tsx",
+                    expoProject,
                     "Android Expo Debug test(localhost)",
-                    expoWorkspace,
                     ExpoLocalDebugConfigName,
                     Platform.AndroidExpo,
                     1,
                 );
             });
-        }
-        if (process.platform === "darwin" && testParameters.RunIosTests) {
-            it("iOS Expo app Debug test(Tunnel)", async function () {
+
+            it("Android Expo app Debug test(Tunnel)", async function () {
                 this.timeout(debugExpoTestTime);
+                if (!isLoggedInExpo()) {
+                    assert.fail(expoLogginErrorMessage);
+                }
                 await expoTest(
-                    "App.tsx",
-                    "iOS Expo Debug test(Tunnel)",
-                    expoWorkspace,
-                    ExpoDebugConfigName,
-                    Platform.iOSExpo,
+                    expoProject,
+                    "Android Expo Debug test(Tunnel)",
+                    ExpoTunnelDebugConfigName,
+                    Platform.AndroidExpo,
                     5,
                 );
             });
+        }
 
+        if (process.platform === "darwin" && testParameters.RunIosTests) {
             it("iOS Pure RN app Expo test(LAN)", async function () {
                 this.timeout(debugExpoTestTime);
                 await expoTest(
-                    "App.js",
+                    pureProject,
                     "iOS pure RN Expo test(LAN)",
-                    pureWorkspace,
                     ExpoLanDebugConfigName,
                     Platform.iOSExpo,
                     1,
@@ -332,9 +348,8 @@ export function startExpoTests(
             it("iOS Expo app Debug test(LAN)", async function () {
                 this.timeout(debugExpoTestTime);
                 await expoTest(
-                    "App.tsx",
+                    expoProject,
                     "iOS Expo Debug test(LAN)",
-                    expoWorkspace,
                     ExpoLanDebugConfigName,
                     Platform.iOSExpo,
                     1,
@@ -344,12 +359,25 @@ export function startExpoTests(
             it("iOS Expo app Debug test(localhost)", async function () {
                 this.timeout(debugExpoTestTime);
                 await expoTest(
-                    "App.tsx",
+                    expoProject,
                     "iOS Expo Debug test(localhost)",
-                    expoWorkspace,
                     ExpoLocalDebugConfigName,
                     Platform.iOSExpo,
                     1,
+                );
+            });
+
+            it("iOS Expo app Debug test(Tunnel)", async function () {
+                this.timeout(debugExpoTestTime);
+                if (!isLoggedInExpo()) {
+                    assert.fail(expoLogginErrorMessage);
+                }
+                await expoTest(
+                    expoProject,
+                    "iOS Expo Debug test(Tunnel)",
+                    ExpoTunnelDebugConfigName,
+                    Platform.iOSExpo,
+                    5,
                 );
             });
         }